src/components: move FormLogin global login comps
